+++ src/context/SidebarContext.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useState } from 'react';
+import React, { createContext, useState } from 'react';
 
 type SidebarContextType = {
   selectedTab: string;
@@ -16,11 +16,4 @@ export const SidebarProvider = ({ children }: { children: React.ReactNode }) => 
     </SidebarContext.Provider>
   );
 };
-
-export const useSidebar = () => {
-  const context = useContext(SidebarContext);
-  if (!context) {
-    throw new Error('useSidebar must be used within a SidebarProvider');
-  }
-  return context;
-};
+export { SidebarContext };
