+++ src/features/hr/Users.tsx
@@ -3,7 +3,7 @@ import Users from "../../mocks/dummy_users_1000";
 import { useNavigate } from "react-router-dom";
 
 export const UsersPage = () => {
-  const [hrId, setHrId] = useState("600b18f1-bdef-4ec3-8819-96a84058f786");
+  const [hrId] = useState("600b18f1-bdef-4ec3-8819-96a84058f786");
   const [searchTerm, setSearchTerm] = useState("");
   const navigate = useNavigate();
   const filteredUsers = Users.filter(
